--- docs/docs/auto-docs/plugin/available/plugin_map/variables/ExtensionPointsDashboard.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: ExtensionPointsDashboard
-
-> `const` **ExtensionPointsDashboard**: `React.FC`
-
-Defined in: src/plugin/available/plugin\_map/pages/ExtensionPointsDashboard.tsx:33
