--- docs/docs/auto-docs/plugin/available/plugin_map/variables/ExtensionPointsGlobal.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: ExtensionPointsGlobal
-
-> `const` **ExtensionPointsGlobal**: `React.FC`
-
-Defined in: src/plugin/available/plugin\_map/pages/ExtensionPointsGlobal.tsx:46
