--- docs/docs/auto-docs/plugin/available/plugin_map/variables/ExtensionPointsOrganization.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: ExtensionPointsOrganization
-
-> `const` **ExtensionPointsOrganization**: `React.FC`
-
-Defined in: src/plugin/available/plugin\_map/pages/ExtensionPointsOrganization.tsx:18
