--- docs/docs/auto-docs/plugin/available/plugin_map/variables/ExtensionPointsUser.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: ExtensionPointsUser
-
-> `const` **ExtensionPointsUser**: `React.FC`
-
-Defined in: src/plugin/available/plugin\_map/pages/ExtensionPointsUser.tsx:18
