--- docs/docs/auto-docs/plugin/available/plugin_map/variables/MapIconInjector.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: MapIconInjector
-
-> `const` **MapIconInjector**: `React.FC`\<`MapIconInjectorProps`\>
-
-Defined in: src/plugin/available/plugin\_map/injector/MapIconInjector.tsx:19
