--- docs/docs/auto-docs/plugin/available/plugin_map/variables/default.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: default
-
-> `const` **default**: [`IPluginLifecycle`](../../../types/interfaces/IPluginLifecycle.md)
-
-Defined in: src/plugin/available/plugin\_map/index.tsx:22
