+++ docs/docs/auto-docs/plugin/registry/functions/discoverAndRegisterAllPlugins.md
@@ -6,7 +6,7 @@
 
 > **discoverAndRegisterAllPlugins**(): `Promise`\<`void`\>
 
-Defined in: [src/plugin/registry.tsx:267](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/registry.tsx#L267)
+Defined in: [src/plugin/registry.tsx:234](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/registry.tsx#L234)
 
 Discover and register all plugins from plugin manager
