+++ docs/docs/auto-docs/plugin/registry/functions/getPluginComponents.md
@@ -6,7 +6,7 @@
 
 > **getPluginComponents**(`pluginId`): `Record`\<`string`, `ComponentType`\<\{\}\>\>
 
-Defined in: [src/plugin/registry.tsx:298](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/registry.tsx#L298)
+Defined in: [src/plugin/registry.tsx:265](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/registry.tsx#L265)
 
 Get all components for a plugin
