+++ docs/docs/auto-docs/plugin/registry/functions/initializePluginSystem.md
@@ -6,7 +6,7 @@
 
 > **initializePluginSystem**(): `Promise`\<`void`\>
 
-Defined in: [src/plugin/registry.tsx:324](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/registry.tsx#L324)
+Defined in: [src/plugin/registry.tsx:291](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/registry.tsx#L291)
 
 Initialize the plugin system (call this on app startup)
