+++ docs/docs/auto-docs/plugin/services/AdminPluginFileService/variables/adminPluginFileService.md
@@ -6,6 +6,6 @@
 
 > `const` **adminPluginFileService**: [`AdminPluginFileService`](../classes/AdminPluginFileService.md)
 
-Defined in: [src/plugin/services/AdminPluginFileService.ts:458](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/services/AdminPluginFileService.ts#L458)
+Defined in: [src/plugin/services/AdminPluginFileService.ts:457](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/services/AdminPluginFileService.ts#L457)
 
 Singleton instance export
