+++ docs/docs/auto-docs/plugin/variables/PluginRouteRenderer.md
@@ -6,6 +6,6 @@
 
 > `const` **PluginRouteRenderer**: `React.FC`\<`PluginRouteRendererProps`\>
 
-Defined in: [src/plugin/PluginRouteRenderer.tsx:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/PluginRouteRenderer.tsx#L20)
+Defined in: [src/plugin/routes/PluginRouteRenderer.tsx:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/routes/PluginRouteRenderer.tsx#L20)
 
 Component to render plugin routes using the plugin registry
