+++ docs/docs/auto-docs/plugin/variables/PluginRoutes.md
@@ -6,6 +6,6 @@
 
 > `const` **PluginRoutes**: `React.FC`\<`PluginRoutesProps`\>
 
-Defined in: [src/plugin/PluginRoutes.tsx:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/PluginRoutes.tsx#L22)
+Defined in: [src/plugin/routes/PluginRoutes.tsx:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/plugin/routes/PluginRoutes.tsx#L22)
 
 Component that renders plugin routes dynamically
