+++ src/plugin/index.ts
@@ -7,8 +7,8 @@
 
 export { default as PluginManager } from './manager';
 
-export { default as PluginRoutes } from './PluginRoutes';
-export { default as PluginRouteRenderer } from './PluginRouteRenderer';
+export { default as PluginRoutes } from './routes/PluginRoutes';
+export { default as PluginRouteRenderer } from './routes/PluginRouteRenderer';
 
 export {
   usePluginRoutes,
