+++ src/plugin/registry.tsx
@@ -140,9 +140,6 @@ async function getPluginManifest(
 function extractComponentNames(manifest: IPluginManifest): Set<string> {
   const componentNames = new Set<string>();
 
-  console.log('=== EXTRACTING COMPONENT NAMES ===');
-  console.log('Plugin manifest:', manifest.pluginId);
-
   // Handle all route types
   const routeArrays = [
     manifest.extensionPoints?.routes,
@@ -154,10 +151,7 @@ function extractComponentNames(manifest: IPluginManifest): Set<string> {
 
   routeArrays.forEach((routes, index) => {
     if (routes) {
-      const routeType = ['routes', 'RA1', 'RA2', 'RU1', 'RU2'][index];
-      console.log(`${routeType} found:`, routes.length);
       routes.forEach((route: IRouteExtension) => {
-        console.log('Route:', route.path, 'Component:', route.component);
         if (route.component) {
           componentNames.add(route.component);
         }
@@ -176,8 +170,7 @@ function extractComponentNames(manifest: IPluginManifest): Set<string> {
 
   drawerArrays.forEach((drawer, index) => {
     if (drawer) {
-      const drawerType = ['drawer', 'DA1', 'DA2', 'DU1', 'DU2'][index];
-      console.log(`${drawerType} found:`, drawer.length);
+      // Drawer components don't need to be added to componentNames
     }
   });
 
@@ -190,15 +183,7 @@ function extractComponentNames(manifest: IPluginManifest): Set<string> {
 
   injectorArrays.forEach((injectors, index) => {
     if (injectors) {
-      const injectorType = ['G1', 'G2', 'G3'][index];
-      console.log(`${injectorType} found:`, injectors.length);
       injectors.forEach((injector: IInjectorExtension) => {
-        console.log(
-          'Injector:',
-          injector.injector,
-          'Description:',
-          injector.description,
-        );
         if (injector.injector) {
           componentNames.add(injector.injector);
         }
@@ -206,8 +191,6 @@ function extractComponentNames(manifest: IPluginManifest): Set<string> {
     }
   });
 
-  console.log('=== COMPONENT NAMES EXTRACTED ===');
-  console.log('Component names:', Array.from(componentNames));
   return componentNames;
 }
 
@@ -218,10 +201,7 @@ export async function registerPluginDynamically(
   pluginId: string,
 ): Promise<void> {
   try {
-    console.log(`=== REGISTERING PLUGIN: ${pluginId} ===`);
-
     if (pluginRegistry[pluginId]) {
-      console.log(`Plugin ${pluginId} already registered, skipping`);
       return;
     }
 
@@ -231,31 +211,18 @@ export async function registerPluginDynamically(
       return;
     }
 
-    console.log(`Plugin status: ${loadedPlugin.status}`);
     if (loadedPlugin.status !== 'active') {
-      console.log(`Plugin ${pluginId} is not active, skipping`);
       return;
     }
 
     // Use the components that are already loaded by the plugin manager
     const components = loadedPlugin.components;
 
-    console.log(
-      `Components from plugin manager:`,
-      Object.keys(components || {}),
-    );
-
     if (components && Object.keys(components).length > 0) {
       pluginRegistry[pluginId] = components;
-      console.log(
-        `Plugin ${pluginId} registered successfully with components:`,
-        Object.keys(components),
-      );
     } else {
       console.warn(`No components found for plugin ${pluginId}`);
     }
-
-    console.log(`=== END REGISTERING PLUGIN: ${pluginId} ===`);
   } catch (error) {
     console.error(`Failed to register plugin '${pluginId}':`, error);
   }
