+++ src/plugin/tests/routes/PluginRouteRenderer.spec.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import PluginRouteRenderer from '../PluginRouteRenderer';
-import { getPluginComponents, isPluginRegistered } from '../registry';
+import PluginRouteRenderer from '../../routes/PluginRouteRenderer';
+import { getPluginComponents, isPluginRegistered } from '../../registry';
 
 // Mock the registry
 vi.mock('../registry', () => ({
